+++ react_on_rails_pro/spec/dummy/client/app/ror-auto-load-components/AsyncComponentsTreeForTesting.jsx
@@ -3,6 +3,8 @@ import HelloWorldHooks from './HelloWorldHooks';
 
 const AsyncComponentsBranch = ({ branchName, level }) => {
   const buildResult = () => {
+    console.log(`${branchName} (level ${level})`);
+    console.error('Error message', { branchName, level });
     if (level === 0) {
       return <div>{`${branchName} (level 0)`}</div>;
     }
@@ -40,6 +42,7 @@ const AsyncHelloWorldHooks = (props) => {
 };
 
 const AsyncComponentsTreeForTesting = (props) => {
+  console.log('Sync console log from AsyncComponentsTreeForTesting');
   return (
     <div>
       <div>