+++ react_on_rails_pro/spec/dummy/spec/helpers/react_on_rails_pro_helper_spec.rb
@@ -187,9 +187,15 @@ describe ReactOnRailsProHelper, type: :helper do
     let(:component_options) { { prerender: true, trace: true, id: "#{component_name}-react-component-0" } }
     let(:chunks) do
       [
-        "<div>Chunk 1: Stream React Server Components</div>",
-        "<div>Chunk 2: More content</div>",
-        "<div>Chunk 3: Final content</div>"
+        { html: "<div>Chunk 1: Stream React Server Components</div>",
+          consoleReplayScript: "<script>console.log.apply(console, " \
+                               "['Chunk 1: Console Message'])</script>" },
+        { html: "<div>Chunk 2: More content</div>",
+          consoleReplayScript: "<script>console.log.apply(console, " \
+                               "['Chunk 2: Console Message']);\n" \
+                               "console.error.apply(console, " \
+                               "['Chunk 2: Console Error']);</script>" },
+        { html: "<div>Chunk 3: Final content</div>", consoleReplayScript: "" }
       ]
     end
     let(:chunks_read) { [] }
@@ -205,7 +211,7 @@ describe ReactOnRailsProHelper, type: :helper do
     end
     let(:react_component_div_with_initial_chunk) do
       <<-HTML.strip
-        <div id="StreamAsyncComponents-react-component-0">#{chunks.first}</div>
+        <div id="StreamAsyncComponents-react-component-0">#{chunks.first[:html]}</div>
       HTML
     end
 
@@ -216,7 +222,7 @@ describe ReactOnRailsProHelper, type: :helper do
         allow(response).to receive(:read_body) do |&read_body_block|
           chunks.each do |chunk|
             chunks_read << chunk
-            read_body_block.call(chunk)
+            read_body_block.call(chunk.to_json)
           end
         end
         block.call(response)
@@ -235,9 +241,10 @@ describe ReactOnRailsProHelper, type: :helper do
         mock_request_and_response
       end
 
-      it "returns the component shell that exist in the initial chunk" do
+      it "returns the component shell that exist in the initial chunk with the consoleReplayScript" do
         initial_result = stream_react_component(component_name, props: props, **component_options)
         expect(initial_result).to include(react_component_div_with_initial_chunk)
+        expect(initial_result).to include(chunks.first[:consoleReplayScript])
         expect(initial_result).not_to include("More content", "Final content")
         expect(chunks_read.count).to eq(1)
       end
@@ -249,12 +256,15 @@ describe ReactOnRailsProHelper, type: :helper do
         expect(fiber).to be_alive
 
         second_result = fiber.resume
-        expect(second_result).to eq(chunks[1])
+        # regex that matches the html and consoleReplayScript and allows for any amount of whitespace between them
+        expect(second_result).to match(
+          /#{Regexp.escape(chunks[1][:html])}\s+#{Regexp.escape(chunks[1][:consoleReplayScript])}/
+        )
         expect(second_result).not_to include("Stream React Server Components", "Final content")
         expect(chunks_read.count).to eq(2)
 
         third_result = fiber.resume
-        expect(third_result).to eq(chunks[2])
+        expect(third_result).to eq(chunks[2][:html].to_s)
         expect(third_result).not_to include("Stream React Server Components", "More content")
         expect(chunks_read.count).to eq(3)
 
@@ -310,7 +320,7 @@ describe ReactOnRailsProHelper, type: :helper do
 
         # Check that the Rails context is before the first chunk
         rails_context_index = initial_result.index('id="js-react-on-rails-context"')
-        first_chunk_index = initial_result.index(chunks.first)
+        first_chunk_index = initial_result.index(chunks.first[:html])
         expect(rails_context_index).to be < first_chunk_index
 
         # The following chunks should not include the Rails context