+++ react_on_rails_pro/spec/dummy/spec/rails_helper.rb
@@ -81,12 +81,42 @@ RSpec.configure do |config|
   # selenium_firefox webdriver only works for Travis-CI builds.
   default_driver = :selenium_chrome_headless
 
-  supported_drivers = %i[selenium_chrome_headless selenium_chrome selenium selenium_headless]
+  # This code is copied from Capybara source code
+  # https://github.com/teamcapybara/capybara/blob/0480f90168a40780d1398c75031a255c1819dce8/lib/capybara/registrations/drivers.rb#L20-L42
+  # It uses the same code to define chrome driver but adds the logging prefs
+  Capybara.register_driver :selenium_chrome_with_logging do |app|
+    version = Capybara::Selenium::Driver.load_selenium
+    options_key = Capybara::Selenium::Driver::CAPS_VERSION.satisfied_by?(version) ? :capabilities : :options
+    browser_options = Selenium::WebDriver::Chrome::Options.new.tap do |opts|
+      # Workaround https://bugs.chromium.org/p/chromedriver/issues/detail?id=2650&q=load&sort=-id&colspec=ID%20Status%20Pri%20Owner%20Summary
+      opts.add_argument("--disable-site-isolation-trials")
+      opts.add_option("goog:loggingPrefs", { browser: "ALL" })
+    end
+
+    Capybara::Selenium::Driver.new(app, **{ :browser => :chrome, options_key => browser_options })
+  end
+
+  Capybara.register_driver :selenium_chrome_headless_with_logging do |app|
+    version = Capybara::Selenium::Driver.load_selenium
+    options_key = Capybara::Selenium::Driver::CAPS_VERSION.satisfied_by?(version) ? :capabilities : :options
+    browser_options = Selenium::WebDriver::Chrome::Options.new.tap do |opts|
+      opts.add_argument("--headless=new")
+      opts.add_argument("--disable-gpu") if Gem.win_platform?
+      # Workaround https://bugs.chromium.org/p/chromedriver/issues/detail?id=2650&q=load&sort=-id&colspec=ID%20Status%20Pri%20Owner%20Summary
+      opts.add_argument("--disable-site-isolation-trials")
+      opts.add_option("goog:loggingPrefs", { browser: "ALL" })
+    end
+
+    Capybara::Selenium::Driver.new(app, **{ :browser => :chrome, options_key => browser_options })
+  end
+
+  supported_drivers = %i[selenium_chrome_headless selenium_chrome]
   driver = ENV["DRIVER"].try(:to_sym).presence || default_driver
+  raise "Unsupported driver: #{driver} (supported = #{supported_drivers})" unless supported_drivers.include?(driver)
+
+  driver = :"#{driver}_with_logging"
   Capybara.javascript_driver = driver
   Capybara.default_driver = driver
-
-  raise "Unsupported driver: #{driver} (supported = #{supported_drivers})" unless supported_drivers.include?(driver)
 
   Capybara.register_server(Capybara.javascript_driver) do |app, port|
     require "rack/handler/puma"